+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The Commands of the bot are its API so here's how it will adhere to Semantic Ver
 **MINOR** - Adding new commands or functionality that won't affect or break existing ones    
 **PATCH** - Bug fixes, code improvements (backwards-compatible bug fixes)
 
+## [1.16.0] - 2017-05-02
+### Added
+- !random - returns a random trigger and its name
+- !search - fuzzy search trigger names
+
 ## [1.15.0] - 2017-02-24
 ### Changed
 - fixed cleverbot... until they decide to change things again :-/